fix compiler wornings

+++ src/edhoc/cert.c
@@ -95,7 +95,7 @@ static int deser_sign_cb(void *void_ctx, int tag, unsigned char *start,
 		uint8_t *unit_end = ctx->seek + ctx->unit_size;
 		if (unit_end <= ctx->end) {
 			memcpy(ctx->seek, start + len - ctx->unit_size,
-			       ctx->unit_size);
+			       (uint32_t)ctx->unit_size);
 			ctx->seek = unit_end;
 		}
 	}
@@ -186,11 +186,13 @@ enum err cert_c509_verify(const uint8_t *cert, uint32_t cert_len,
 
 	/*verify the certificates signature*/
 	TRY(verify((enum sign_alg)c._cert_issuer_signature_algorithm, root_pk,
-		   root_pk_len, cert, cert_len - 2 - c._cert_signature.len,
-		   c._cert_signature.value, c._cert_signature.len, verified));
+		   root_pk_len, cert,
+		   cert_len - 2 - (uint32_t)c._cert_signature.len,
+		   c._cert_signature.value, (uint32_t)c._cert_signature.len,
+		   verified));
 
-	TRY(_memcpy_s(pk, *pk_len, c._cert_pk.value, c._cert_pk.len));
-	*pk_len = c._cert_pk.len;
+	TRY(_memcpy_s(pk, *pk_len, c._cert_pk.value, (uint32_t)c._cert_pk.len));
+	*pk_len = (uint32_t)c._cert_pk.len;
 
 	return ok;
 }
@@ -257,10 +259,10 @@ enum err cert_x509_verify(const uint8_t *cert, uint32_t cert_len,
 		mbedtls_x509_crt_free(&m_cert);
 		return unsupported_signature_algorithm;
 	}
-	size_t hash_len = mbedtls_md_get_size(md_info);
+	int hash_len = mbedtls_md_get_size(md_info);
 
 	size_t sig_len = get_signature_len(sign_alg);
-	TRY(check_buffer_size(SIGNATURE_DEFAULT_SIZE, sig_len));
+	TRY(check_buffer_size(SIGNATURE_DEFAULT_SIZE, (uint32_t)sig_len));
 	uint8_t sig[SIGNATURE_DEFAULT_SIZE];
 
 	/* get the public key of the CA */
@@ -282,8 +284,8 @@ enum err cert_x509_verify(const uint8_t *cert, uint32_t cert_len,
 	}
 
 	/*verify the certificates signature*/
-	TRY(verify(sign_alg, root_pk, root_pk_len, m_cert.tbs.p, m_cert.tbs.len,
-		   sig, sig_len, verified));
+	TRY(verify(sign_alg, root_pk, root_pk_len, m_cert.tbs.p,
+		   (uint32_t)m_cert.tbs.len, sig, (uint32_t)sig_len, verified));
 
 	/* export the public key from certificate */
 	{
@@ -296,10 +298,10 @@ enum err cert_x509_verify(const uint8_t *cert, uint32_t cert_len,
 			if (*cpk == 0) {
 				++cpk;
 			}
-			cpk_len = m_cert.pk_raw.len - (cpk - m_cert.pk_raw.p);
+			cpk_len = m_cert.pk_raw.len - (size_t)(cpk - m_cert.pk_raw.p);
 		}
-		TRY(_memcpy_s(pk, *pk_len, cpk, cpk_len));
-		*pk_len = cpk_len;
+		TRY(_memcpy_s(pk, *pk_len, cpk, (uint32_t)cpk_len));
+		*pk_len = (uint32_t)cpk_len;
 		PRINT_ARRAY("pk from cert", pk, *pk_len);
 	}
